requirements.txt
flask
tensorflow==2.19.0
numpy
pillow
flask-cors
werkzeug
